fix: handling keyboard interrupts
Previously, the camera process did not stop even after a KeyBoardInterrupt.

The monitor() function has been fixed to detect an interrupt during the gstreamer subprocess,
and it now terminates the subprocess shell as well as the OS process without closing the program.
It also logs this info to the terminal.

TODO: better monitoring practices. eg: logging after init maybe